% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_for_mixtcomp.r
\name{addVariable}
\alias{addVariable}
\title{Helper function for Mixtcomp Input formatting}
\usage{
addVariable(lm, data, id, model, paramStr)
}
\arguments{
\item{paramStr}{}
}
\value{

}
\description{
Helper function for Mixtcomp Input formatting
}
