% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_results.r
\name{aggregate_classification_probabilities}
\alias{aggregate_classification_probabilities}
\title{Aggregate the classification probabilities estimated in the nested clustering}
\usage{
aggregate_classification_probabilities(dir, depth)
}
\arguments{
\item{dir}{String, a path to the directory in which are the subclusters to be aggregated}

\item{depth}{maximum depth of the subclustering to plot}

\item{clusters}{vector of the clusters associated with data}
}
\value{
a plotly plot
}
\description{
Aggregate the classification probabilities estimated in the nested clustering
}
\examples{
dir = "data/generated_data"
}
