% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_results.r
\name{aggregate_clusters}
\alias{aggregate_clusters}
\title{Aggregate the hierarchical clusters estimated}
\usage{
aggregate_clusters(dir)
}
\arguments{
\item{dir}{String, a path to the directory in which are the subclusters to be aggregated}
}
\value{
void
}
\description{
Aggregate the hierarchical clusters estimated
}
\examples{
aggregate_clusters(dir_path_result)
}
\seealso{
Other results aggregation: \code{\link{aggregate_classification_probabilities}},
  \code{\link{aggregate_completed_max}},
  \code{\link{aggregate_completed}}
}
