% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_results.r
\name{aggregate_completed_depth}
\alias{aggregate_completed_depth}
\title{Plot functional data by clusters}
\usage{
aggregate_completed_depth(dir, var, depth)
}
\arguments{
\item{var}{The name of the functional data to plot}

\item{depth}{maximum depth of the subclustering to plot}

\item{clusters}{vector of the clusters associated with data}

\item{output}{The output in which to find the data to plot}

\item{max_nb_lines}{Defines the maximum number of lines to plot (default is 100)}
}
\value{
a plotly plot
}
\description{
Plot functional data by clusters
}
