% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_results.r
\name{aggregate_completed_max}
\alias{aggregate_completed_max}
\title{Aggregate the hierarchical clusters completed data}
\usage{
aggregate_completed_max(dir, var)
}
\arguments{
\item{dir}{String, a path to the directory in which are the subclusters to be aggregated}

\item{var}{String, The variable to be aggregated}
}
\value{
void
}
\description{
Aggregate the hierarchical clusters completed data
}
\examples{
aggregate_clusters(dir_path_result)
}
\seealso{
Other results aggregation: \code{\link{aggregate_classification_probabilities}},
  \code{\link{aggregate_clusters}},
  \code{\link{aggregate_completed}}
}
