% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_toolbox.r
\name{convertFunctional}
\alias{convertFunctional}
\title{Convert a mixtcomp string into 2-columns matrix}
\usage{
convertFunctional(x)
}
\arguments{
\item{x}{a string containing a fonctional observation (cf example)}
}
\value{
matrix
}
\description{
Convert a mixtcomp string into 2-columns matrix
}
\examples{
\dontrun{
convertFunctional("1:5,1.5:12,1.999:2.9")}
}
