% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtcomp_hierachical_launch.R
\name{create_subdirectories}
\alias{create_subdirectories}
\title{Title}
\usage{
create_subdirectories(dir, nClass)
}
\arguments{
\item{dir}{A directory}

\item{nClass}{A positive integer}
}
\value{
void
}
\description{
Title
}
\examples{
create_subdirectories(getwd(),2)
}
