% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_toolbox.r
\name{createFunctional}
\alias{createFunctional}
\title{créé le format fonctionnel de mixtcomp}
\usage{
createFunctional(temps, valeur)
}
\arguments{
\item{temps}{vector of numeric}

\item{valeur}{vector of numeric}
}
\value{
The functional data formatted to the mixtcomp standard
}
\description{
créé le format fonctionnel de mixtcomp
}
\examples{
\dontrun{
Mat = matrix(c(1,2,3,9,1,1.5,15,1000),ncol=2)
convertFunctional(Mat[,1],Mat[,2])}
}
