% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtcomp_hierachical_launch.r
\name{expand}
\alias{expand}
\title{Launch an expand phase : if the directory doesn't contain any subdirectories,
cut the data in clusters according to the clusters estimated by Mixtcomp and given
by an output contained in the dir parameter, create a set of subclusters, copy the data
into them and launch Mixtcomp in each of them.}
\usage{
expand(dir, nClass, strategy = NULL, mcStrategy = NULL, nCore = 1)
}
\arguments{
\item{dir}{The path of the directory to expand}

\item{nClass}{The number of class for the Mixtcomp launches on the next subclusters}

\item{strategy}{(optional) List of three elements which defines the strategy to follow during the hierarchical clustering.
var (name of the categorical variable on which to base the strategy), threshold_nInd (the minimal number of observations in the dataset under which the subclustering cannot be accepted)
threshold_purity (the maximal purity observed in the dataset above which the subclustering is not necessary). At each hierarchical iterations these elements will be tested.}

\item{mcStrategy}{(optional) Additional parameters for mixtcomp launch (nbBurnInIter, nbIter nbGibbsBurnInIter, nbGibbsIter)}
}
\value{
void
}
\description{
Launch an expand phase : if the directory doesn't contain any subdirectories,
cut the data in clusters according to the clusters estimated by Mixtcomp and given
by an output contained in the dir parameter, create a set of subclusters, copy the data
into them and launch Mixtcomp in each of them.
}
