% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtcomp_hierachical_launch.r
\name{extract_data_per_cluster}
\alias{extract_data_per_cluster}
\title{During the expand phase, cut data according to clusters and copy to subfolder}
\usage{
extract_data_per_cluster(dir)
}
\arguments{
\item{dir}{A String}
}
\value{
void
}
\description{
During the expand phase, cut data according to clusters and copy to subfolder
}
