% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.r
\name{generate_data}
\alias{generate_data}
\title{Generate data for testing}
\usage{
generate_data(size = 1000)
}
\arguments{
\item{size}{integer, number of observatios generated}
}
\value{
void
}
\description{
Generate data for testing
}
