% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtcomp_hierachical_launch.r
\name{launch_Mixtcomp_Hierarchical}
\alias{launch_Mixtcomp_Hierarchical}
\title{Launch the Hierarchical Mixtcomp.}
\usage{
launch_Mixtcomp_Hierarchical(data_path, descriptor_path, nClass, depth,
  output_dir = NULL)
}
\arguments{
\item{data_path}{A string}

\item{descriptor_path}{A string}

\item{nClass}{Integer between 0 and 10}

\item{depth}{Positive integer}

\item{output_dir}{(optional) if not null, the results will be written in this directory.
If Null, a directory with the name of the dataset will be created and used as output directory instead.}
}
\value{
void
}
\description{
Launch the Hierarchical Mixtcomp.
}
\examples{
launch_Mixtcomp_Hierarchical("data/data.csv","data/descriptor.csv",3,3)
\dontrun{
param <- list(
x = c(0., 10., 20.),
y = c(0., 10., 0.),
sd = 0.1,
tMin = 0.,
tMax = 20.,
nTime = 100)

data = c()
for(i in 1:2000){
  data <- c(data,functionalInterPolyGenerator(param=param))
}

data <- cbind(data,rnorm(n = 2000,mean = c(0,5,10,100),sd=c(1,2,3,4)))
colnames(data) = c("F1","X1")
descriptor <-
  setNames(data.frame(matrix(
    ncol = ncol(data),
    nrow = 2,
   byrow = TRUE
 )), colnames(data))
descriptor[1,] <-
 c("Functional","Gaussian_sjk")
descriptor[2, 1] <- rep("nSub: 5, nCoeff: 2", 1)

write.table(
  data,
  "data/data_mixtcomp.csv" ,
  sep = ";",
  na = "",
  row.names = F
)
write.table(descriptor,file="data/descriptor.csv",sep=";",na = "",row.names = F)
}
}
\author{
Etienne Goffinet
}
