% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtcomp_hierachical_launch.r
\name{launch_Mixtcomp_Hierarchical_predict}
\alias{launch_Mixtcomp_Hierarchical_predict}
\title{Launch the Mixtcomp prediction based on an existing hierarchical cluster}
\usage{
launch_Mixtcomp_Hierarchical_predict(data_path, param_dir, output_dir = NULL)
}
\arguments{
\item{data_path}{A string}

\item{output_dir}{(optional) if not null, the results will be written in this directory.
If Null, a directory with the name of the dataset will be created and used as output directory instead.}

\item{descriptor_path}{A string}

\item{nClass}{Integer between 0 and 10}

\item{depth}{Positive integer}
}
\value{
void
}
\description{
Launch the Mixtcomp prediction based on an existing hierarchical cluster
}
\examples{
launch_Mixtcomp_Hierarchical("data/data.csv","data/descriptor.csv",3,3)
}
\author{
Etienne Goffinet
}
