% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_toolbox.r
\name{linesFunctional}
\alias{linesFunctional}
\title{lines functional data}
\usage{
linesFunctional(x, ...)
}
\arguments{
\item{x}{Functional data in mixtcomp string format}

\item{...}{additional plot parameter}
}
\value{
void
}
\description{
lines functional data
}
\examples{
\dontrun{
#' plotFunctional("1:5,3:2,6:9.78,9:4,12:25")
linesFunctional("1:7,3:4,6:11.78,9:5,12:12")
}
}
