% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixtCompRTools-package.R
\docType{package}
\name{MixtCompRTools-package}
\alias{MixtCompRTools-package}
\title{MixtCompRTools}
\description{
Add tools for MixtComp, essentially Hierarchical calls. Allow to launch MixtComp hierarchically, in both learn and predict mode.
}
\keyword{package}
