% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtcomp_hierachical_launch.r
\name{plot_functional_hierarchique}
\alias{plot_functional_hierarchique}
\title{Title}
\usage{
plot_functional_hierarchique(clusters, var, output, depth, max_nb_lines = 100)
}
\arguments{
\item{max_nb_lines}{}
}
\description{
Title
}
