% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{plotFunctional}
\alias{plotFunctional}
\title{Plot functional data}
\usage{
plotFunctional(x, ...)
}
\arguments{
\item{x}{Functional data in mixtcomp string format}

\item{...}{additional plot parameter}
}
\description{
Plot functional data
}
\examples{
plotFunctional("1:5,3:2,6:9.78,9:4,12:25")

}
\author{
Quentin Grimonprez
}
