% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtcomp_hierachical_launch.r
\name{prune_hierarchy}
\alias{prune_hierarchy}
\title{Prune a hierarchical clsutering. The pruning can be done according to three different strategies}
\usage{
prune_hierarchy(dir, purity_strategy = list(var = NULL, threshold =
  NULL), cluster_size_strategy = list(threshold = NULL))
}
\arguments{
\item{dir}{A string, the path of the hierarchical clustering to prune}

\item{purity_strategy}{a list of two elements (var and threshold), that defines which variable purity to check and according to which threshold (double between 0 and 1)}

\item{cluster_size_strategy}{a list of one element (threshold) that define the cluster size threshold.}
}
\description{
- The purity of the clusters : if a cluster is purer than a certain threshold then the subclusters are not useful
 - The size of the cluster : below a certain size, a cluster should not have been subclusterized.
 - The quality of the completion : Is the global completion more accurate with this subclusterization ? Aims at avoiding overfitting.
}
\details{
These strategies can be cumulated.
}
\author{
Etienne Goffinet
}
