% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_for_mixtcomp.r
\name{addVariable}
\alias{addVariable}
\title{Helper function for Mixtcomp Input formatting}
\usage{
addVariable(lm, data, id, model, paramStr)
}
\arguments{
\item{lm}{list as container}

\item{data}{Vector of data}

\item{id}{Name of the variable}

\item{model}{Name of the model : "Gaussian_pjk", ..}

\item{paramStr}{Additional parameters as strings}
}
\value{

}
\description{
Helper function for Mixtcomp Input formatting
}
