% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_results.r
\name{aggregate_classification_probabilities}
\alias{aggregate_classification_probabilities}
\title{Aggregate the classification probabilities estimated in the nested clustering}
\usage{
aggregate_classification_probabilities(dir, depth)
}
\arguments{
\item{dir}{String, path of the directory in which to finc the nested clustering}

\item{depth}{Positive Integer, maximum depth of the subclustering to plot}
}
\value{
a plotly plot
}
\description{
Aggregate the classification probabilities estimated in the nested clustering
}
\examples{
dir = "data/generated_data"
}
\seealso{
Other results aggregation: \code{\link{aggregate_clusters}},
  \code{\link{aggregate_completed_max}},
  \code{\link{aggregate_completed}}
}
