% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_results.r
\name{aggregate_completed}
\alias{aggregate_completed}
\title{Aggregate the hierarchical clusters completed data}
\usage{
aggregate_completed(dir, var)
}
\arguments{
\item{dir}{String, a path to the directory in which are the subclusters to be aggregated}

\item{var}{The variable to be aggregated}
}
\value{
void
}
\description{
Aggregate the hierarchical clusters completed data
}
\examples{
aggregate_clusters(dir_path_result)
}
