% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_results.r
\name{aggregate_completed}
\alias{aggregate_completed}
\title{Plot functional data by clusters}
\usage{
aggregate_completed(dir, var, depth = NULL)
}
\arguments{
\item{dir}{String, path of the directory in which to finc the nested clustering}

\item{var}{String, The name of the functional data to plot}

\item{depth}{Positive Integer or NULL, maximum depth of the subclustering to plot. If NULL, the maximal depth of the clusters is assumed.}
}
\value{
a plotly plot
}
\description{
Plot functional data by clusters
}
\seealso{
Other results aggregation: \code{\link{aggregate_classification_probabilities}},
  \code{\link{aggregate_clusters}},
  \code{\link{aggregate_completed_max}}
}
