% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtcomp_hierachical_launch.r
\name{create_subdirectories}
\alias{create_subdirectories}
\title{Create Subdirectories during the expand phase}
\usage{
create_subdirectories(dir, nClass)
}
\arguments{
\item{dir}{A directory}

\item{nClass}{A positive integer}
}
\value{
void
}
\description{
Create Subdirectories during the expand phase
}
