% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtcomp_hierachical_launch.r
\name{expand}
\alias{expand}
\title{Launch an expand phase : if the directory doesn't contain any subdirectories,
cut the data in clusters according to the clusters estimated by Mixtcomp and given
by an output contained in the dir parameter, create a set of subclusters, copy the data
into them and launch Mixtcomp in each of them.}
\usage{
expand(dir, nClass)
}
\arguments{
\item{dir}{The path of the directory to expand}

\item{nClass}{The number of class for the Mixtcomp launches on the next subclusters}
}
\value{
void
}
\description{
Launch an expand phase : if the directory doesn't contain any subdirectories,
cut the data in clusters according to the clusters estimated by Mixtcomp and given
by an output contained in the dir parameter, create a set of subclusters, copy the data
into them and launch Mixtcomp in each of them.
}
\examples{
expand()
}
