% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_for_mixtcomp.r
\name{getData}
\alias{getData}
\title{Transform data into Mixtcomp accepted format}
\usage{
getData(...)
}
\arguments{
\item{...}{List of arguments (usually equal to c(data_path, descriptor_path))}
}
\value{
list of data
}
\description{
Transform data into Mixtcomp accepted format
}
\author{
Vincent Kubicki
}
