% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalMixtComp.R
\name{hierarchicalMixtCompLearn}
\alias{hierarchicalMixtCompLearn}
\title{Hierarchical version of mixtCompLearn}
\usage{
hierarchicalMixtCompLearn(dataList, mcStrategy = list(nbBurnInIter = 100,
  nbIter = 100, nbGibbsBurnInIter = 50, nbGibbsIter = 50, nInitPerClass =
  10, nSemTry = 20), nbClass, confidenceLevel = 0.95,
  criterion = c("BIC", "ICL", "lnObservedLikelihood",
  "lnCompletedLikelihood"), minClassSize = 5)
}
\arguments{
\item{dataList}{\emph{lm} element of the output of \link{getData}}

\item{mcStrategy}{a list containing the parameters of the SEM-Gibbs algorithm}

\item{nbClass}{the number of class of the mixture model}

\item{confidenceLevel}{quantile for confidence interval of estimated parameters.}

\item{criterion}{criterion used to choose the order of clusters: "BIC", "ICL", "lnObservedLikelihood" or "lnCompletedLikelihood"}

\item{minClassSize}{minimum number of elements in a cluster. If a cluster has less elements, this classes won't be splitted anymore.}
}
\value{
a list containing res and history. res is a list containing the mixtComp outputs for 2 clusters to nbClass clusters. History contains intermediate results.
}
\description{
Hierarchical version of mixtCompLearn
}
\details{
1) cluster data in 2 groups
2) A clustering in 2 groups is performed on each groups. The best clustering according to the specified criterion is kept to create a model with 3 groups
3) ...
}
\examples{
\donttest{
set.seed(42)
data <- matrix(nrow = 200, ncol = 2)
data[,1] = c(rnorm(50, 0, 0.5), rnorm(50, -4, 0.8), rnorm(30, 2, 0.5), rnorm(70, -1, 0.4))
data[,2] = c(rnorm(100, -2, 1), rnorm(100, 3, 0.4))
colnames(data) = c("gauss1", "gauss2")
truePartition <- rep(1:4, c(50, 50, 30, 70))

descriptor <- matrix("Gaussian_sjk", nrow = 1, ncol = 2, dimnames = list(NULL, colnames(data)))

resGetData <- getData(list(data, descriptor))

mcStrategy <- list(nbBurnInIter = 100,
                   nbIter = 100,
                   nbGibbsBurnInIter = 50,
                   nbGibbsIter = 50,
                   nInitPerClass = 10,
                   nSemTry = 20)

res <- hierarchicalMixtCompLearn(resGetData$lm, mcStrategy, nbClass = 4)
}

}
\author{
Quentin Grimonprez
}
