% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtcomp_hierachical_launch.r
\name{launch_Mixtcomp_Hierarchical}
\alias{launch_Mixtcomp_Hierarchical}
\title{Launch the Hierarchical Mixtcomp.}
\usage{
launch_Mixtcomp_Hierarchical(data_path, descriptor_path, nClass, depth,
  output_dir = NULL)
}
\arguments{
\item{data_path}{A string}

\item{descriptor_path}{A string}

\item{nClass}{Integer between 0 and 10}

\item{depth}{Positive integer}

\item{output_dir}{(optional) if not null, the results will be written in this directory.
If Null, a directory with the dame of the dataset will be created and used as output directory instead.}
}
\value{
void
}
\description{
Launch the Hierarchical Mixtcomp.
}
\examples{
launch_Mixtcomp_Hierarchical("data/data.csv","data/descriptor.csv",3,3)
}
