% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtcomp_hierachical_launch.r
\name{launch_mixtcomp_predict}
\alias{launch_mixtcomp_predict}
\title{Launch Mixtcomp on a data_mixtcomp.csv and descriptor.csv files contained in
the directory given by the dir param and with nClass number of clusters}
\usage{
launch_mixtcomp_predict(param_dir, test_dir, nClass, mcStrategy)
}
\arguments{
\item{param_dir}{A string defining the path of the directory in which to fint the parameters hierarchical tree on which the hierarchical predict will be based}

\item{test_dir}{A String}

\item{nClass}{Positive intege}
}
\value{
void
}
\description{
Launch Mixtcomp on a data_mixtcomp.csv and descriptor.csv files contained in
the directory given by the dir param and with nClass number of clusters
}
