% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtcomp_hierachical_launch.r
\name{plot_categorical_hierarchique}
\alias{plot_categorical_hierarchique}
\title{Plot categorical data by clusters}
\usage{
plot_categorical_hierarchique(clusters, var, data, depth, max_nb_lines = 100,
  dictionary = NULL, order_by = NULL)
}
\arguments{
\item{clusters}{vector of the clusters associated with data}

\item{var}{The name of the categorical data to plot}

\item{depth}{maximum depth of the subclustering to plot}

\item{max_nb_lines}{Defines the maximum number of lines to plot (default is 100)}

\item{dictionary}{the dictionary in which are stored the real names of the categorical data levels}

\item{output}{The output in which to find the data to plot}
}
\value{
a plotly plot
}
\description{
Plot categorical data by clusters
}
