% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.r
\name{plot_functional_hierarchique}
\alias{plot_functional_hierarchique}
\title{Plot functional data by clusters}
\usage{
plot_functional_hierarchique(clusters, var, data_func, depth,
  max_nb_lines = 100)
}
\arguments{
\item{clusters}{vector of the clusters associated with data}

\item{var}{String, The name of the functional data to plot}

\item{data_func}{list of two elements :time and data / Functional data extracted from Mixtcomp Output}

\item{depth}{Positive Integer, maximum depth of the subclustering to plot}

\item{max_nb_lines}{Positive integer, Defines the maximum number of lines to plot (default is 100)}
}
\value{
a plotly plot
}
\description{
Plot functional data by clusters
}
\seealso{
Other visualisation: \code{\link{plot_categorical_hierarchique}},
  \code{\link{plot_proportion_hierarchique}},
  \code{\link{plot_quality_pred_hierarchique}}
}
