% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.r
\name{plot_proportion_hierarchique}
\alias{plot_proportion_hierarchique}
\title{Plot proportions by clusters}
\usage{
plot_proportion_hierarchique(clusters, var, data, depth, order_by = NULL)
}
\arguments{
\item{clusters}{vector of the clusters associated with data}

\item{var}{String, The name of the categorical data to plot}

\item{data}{the data to be plotted}

\item{depth}{Positive Integer, maximum depth of the subclustering to plot}

\item{order_by}{(optional) element among : size or purity, which define the order of the different cluster in the histogram.}
}
\value{
a plotly plot
}
\description{
Plot proportions by clusters
}
\seealso{
Other visualisation: \code{\link{plot_categorical_hierarchique}},
  \code{\link{plot_functional_hierarchique}},
  \code{\link{plot_quality_pred_hierarchique}}
}
