% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.r
\name{plot_proportion_hierarchique}
\alias{plot_proportion_hierarchique}
\title{Plot proportions by clusters}
\usage{
plot_proportion_hierarchique(clusters, var, data, depth, order_by = NULL)
}
\arguments{
\item{clusters}{vector of the clusters associated with data}

\item{var}{String, The name of the categorical data to plot}

\item{depth}{Positive Integer, maximum depth of the subclustering to plot}

\item{output}{The output in which to find the data to plot}

\item{max_nb_lines}{Positive Integer, Defines the maximum number of lines to plot (default is 100)}

\item{dictionary}{List, the dictionary in which are stored the real names of the categorical data levels}
}
\value{
a plotly plot
}
\description{
Plot proportions by clusters
}
\seealso{
Other visualisation: \code{\link{plot_categorical_hierarchique}},
  \code{\link{plot_functional_hierarchique}},
  \code{\link{plot_quality_pred_hierarchique}}
}
