% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.r
\name{plot_quality_pred_hierarchique}
\alias{plot_quality_pred_hierarchique}
\title{Plot quality of prediction by clusters}
\usage{
plot_quality_pred_hierarchique(clusters, var, idx_test, data_observed,
  data_completed, depth, order_by = NULL)
}
\arguments{
\item{clusters}{vector of the clusters associated with data}

\item{var}{String, The name of the categorical data to plot}

\item{idx_test}{vector of the index of the test data in the data_observed and data_completed}

\item{data_observed}{Observed data to be compared with data_completed}

\item{data_completed}{Completed data}

\item{depth}{Positive Integer, maximum depth of the subclustering to plot}

\item{order_by}{(optional) element among : size, purity or quality, which define the order of the different cluster in the histogram.}
}
\value{
a plotly plot
}
\description{
Plot quality of prediction by clusters
}
\seealso{
Other visualisation: \code{\link{plot_categorical_hierarchique}},
  \code{\link{plot_functional_hierarchique}},
  \code{\link{plot_proportion_hierarchique}}
}
\concept{visualisation}
